requirements/base.txt


# Base
pytz==2021.1
python-slugify==5.0.2
Pillow==8.3.1

# Environment
django-environ==0.4.5

# JWT
pyjwt==2.1.0

#Django
Django==3.2.4

# Cors
django-cors-headers==3.7.0

# Django REST framework
djangorestframework==3.12.4
django-filter==2.4.0

# Passwords security
argon2-cffi==20.1.0

# Static files
whitenoise==5.3.0



